Made the code much more modular.

+++ ftc_app-3.5/TeamCode/src/main/java/org/firstinspires/ftc/teamcode/opmodes/competition/BlueBackAutonomous.kt
@@ -2,13 +2,6 @@ package org.firstinspires.ftc.teamcode.opmodes.competition
 
 import com.qualcomm.robotcore.eventloop.opmode.Autonomous
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode
-import com.qualcomm.robotcore.util.ElapsedTime
-import org.firstinspires.ftc.robotcore.external.navigation.RelicRecoveryVuMark
-import org.firstinspires.ftc.teamcode.libraries.components.lift.RelicRecoveryRobotLift
-import org.firstinspires.ftc.teamcode.libraries.vision.JewelConfigurationDetector
-import org.firstinspires.ftc.teamcode.libraries.vision.PictographIdentifier
-import org.firstinspires.ftc.teamcode.robots.RelicRecoveryRobot
-import kotlin.concurrent.thread
 
 @Autonomous(name = "2: Blue Back", group = "Blue Manual Selection Autonomous")
 class BlueBackAutonomous : LinearOpMode() {
@@ -20,6 +13,7 @@ class BlueBackAutonomous : LinearOpMode() {
     @Throws(InterruptedException::class)
     override fun runOpMode() {
 
+        /*
         // Setup the robot.
         val robot = RelicRecoveryRobot()
         robot.prepareForAutonomous(this)
@@ -28,21 +22,25 @@ class BlueBackAutonomous : LinearOpMode() {
         // Wait for the opmode to start.
         waitForStart()
 
+        // Start the robot.
+        robot.startAuto()
+
+        // Grab the glyph.
+        val glyphGrabbingThread = thread(true) {
+            robot.closeGlyphGrabbers(1100)
+            thread(true) {
+                robot.lift.setPosition(RelicRecoveryRobotLift.LiftPosition.SECOND_LEVEL)
+            }
+        }
+
         // Start the timer.
         val elapsedTime = ElapsedTime(ElapsedTime.Resolution.MILLISECONDS)
-
-        // Start the robot.
-        robot.start()
-
-        thread(start = true) {
-            robot.closeGlyphGrabbers(1200)
-            robot.lift.setPosition(RelicRecoveryRobotLift.LiftPosition.SECOND_LEVEL)
-        }
 
         // Find the position of the jewels.
         val jewelPosition = robot.jewelConfigurationDetector.waitForJewelIdentification(elapsedTime, this)
         robot.jewelConfigurationDetector.disable()
 
+        // Lower the jewel stick as soon as we determine the jewel configuration.
         if (jewelPosition != JewelConfigurationDetector.JewelConfiguration.UNKNOWN) {
             robot.lowerJewelStick(0)
         }
@@ -52,27 +50,38 @@ class BlueBackAutonomous : LinearOpMode() {
         val pictograph = pictographIdentifier.waitForPictographIdentification(elapsedTime, this)
         pictographIdentifier.deactivate()
 
+        // Wait for the glyph to be grabbed.
+        // (I find it funny that the robot identifies the jewels *and* pictograph
+        // before it even has time to grab the glyph that is right in front of it)
+        while (glyphGrabbingThread.isAlive) {
+            sleep(10)
+        }
+
         // Knock off the correct jewel.
         when (jewelPosition) {
             JewelConfigurationDetector.JewelConfiguration.RED_BLUE -> {
                 robot.timeDrive(1000, -0.175)
                 robot.raiseJewelStick()
-                robot.driveOnBalancingStone()
+                robot.driveOnBalancingStone(0.40)
                 robot.driveOffBalancingStone(0.175)
+                robot.powerBreak(false)
             }
 
             JewelConfigurationDetector.JewelConfiguration.BLUE_RED -> {
                 robot.driveOffBalancingStone(0.175)
+                robot.powerBreak(false)
                 robot.raiseJewelStick()
             }
 
             JewelConfigurationDetector.JewelConfiguration.UNKNOWN -> {
                 robot.driveOffBalancingStone(0.175)
+                robot.powerBreak(false)
             }
         }
 
-        // Approach the crypto box.
-        robot.turn(0.45, 90.0)
+        thread(true) {
+            robot.lift.setPosition(RelicRecoveryRobotLift.LiftPosition.FIRST_LEVEL, 0.3)
+        }
 
         // Determine the distance from the wall to the correct crypto box.
         val rightWallDistance = when (pictograph) {
@@ -83,6 +92,7 @@ class BlueBackAutonomous : LinearOpMode() {
         }
 
         // Drive to the correct crypto box.
+        robot.turn(0.3, 90.0)
         robot.driveToDistanceFromLeftObject(rightWallDistance)
 
         // Place the glyph in the correct crypto box.
@@ -92,32 +102,37 @@ class BlueBackAutonomous : LinearOpMode() {
         robot.openGlyphGrabbers(250)
 
         // Back away from the crypto box.
-        robot.timeDrive(850, -0.175)
+        robot.timeDrive(1250, -0.225)
         robot.liftGlyphDeployer(500)
-
-        // Turn towards the center glyphs.
-        robot.turn(0.50, -90.0)
 
         // Open the glyph grabbers.
         robot.releaseGlyphGrabbers()
         robot.retractGlyphDeployer()
 
+        robot.turn(0.20, -90.0)
+
         // Drive to the glyphs in the center and grab one.
-        robot.timeDrive(2500)
+        sleep(1000)
+        robot.timeDrive(475, 1.0)
         robot.closeGlyphGrabbers(1250)
 
         // Drive back the crypto boxes.
-        robot.timeDrive(750, -0.25)
+        robot.timeDrive(650, -0.25)
         thread(true) { robot.lift.setPosition(RelicRecoveryRobotLift.LiftPosition.FIRST_LEVEL) }
-        robot.turn(0.75, 90.0)
+        robot.turn(0.35, 90.0)
         robot.timeDrive(500)
         robot.driveToDistanceFromForwardObject(RelicRecoveryRobot.CRYPTO_BOX_SPACING)
 
         // Line up with the center crypto box.
-        robot.driveToDistanceFromLeftObject(RelicRecoveryRobot.TRAILING_FRONT_CRYPTO_BOX_DISTANCE)
+        robot.driveToDistanceFromLeftObject(RelicRecoveryRobot.CENTER_SIDE_CRYPTO_BOX_DISTANCE)
 
         // Place the glyph in the correct crypto box.
-        robot.lift.drop()
+        if (pictograph == RelicRecoveryVuMark.CENTER) {
+            robot.lift.setPosition(RelicRecoveryRobotLift.LiftPosition.SECOND_LEVEL)
+        } else {
+            robot.lift.drop()
+        }
+
         robot.timeDrive(1000)
         robot.extendGlyphDeployer()
         robot.openGlyphGrabbers(250)
@@ -127,7 +142,8 @@ class BlueBackAutonomous : LinearOpMode() {
         robot.liftGlyphDeployer(500)
 
         // Turn towards the center glyphs.
-        robot.turn(0.25, -90.0)
+        robot.turn(0.35, -90.0)
+        */
     }
 
 }
