+++ ftc_app-3.5/TeamCode/src/main/java/org/firstinspires/ftc/teamcode/opmodes/competition/RedFrontAutonomous.kt
@@ -2,13 +2,6 @@ package org.firstinspires.ftc.teamcode.opmodes.competition
 
 import com.qualcomm.robotcore.eventloop.opmode.Autonomous
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode
-import com.qualcomm.robotcore.util.ElapsedTime
-import org.firstinspires.ftc.robotcore.external.navigation.RelicRecoveryVuMark
-import org.firstinspires.ftc.teamcode.libraries.components.lift.RelicRecoveryRobotLift
-import org.firstinspires.ftc.teamcode.libraries.vision.JewelConfigurationDetector
-import org.firstinspires.ftc.teamcode.libraries.vision.PictographIdentifier
-import org.firstinspires.ftc.teamcode.robots.RelicRecoveryRobot
-import kotlin.concurrent.thread
 
 @Autonomous(name = "3: Red Front", group = "Red Manual Selection Autonomous")
 class RedFrontAutonomous : LinearOpMode() {
@@ -20,6 +13,7 @@ class RedFrontAutonomous : LinearOpMode() {
     @Throws(InterruptedException::class)
     override fun runOpMode() {
 
+        /*
         // Setup the robot.
         val robot = RelicRecoveryRobot()
         robot.prepareForAutonomous(this)
@@ -28,19 +22,25 @@ class RedFrontAutonomous : LinearOpMode() {
         // Wait for the opmode to start.
         waitForStart()
 
+        // Start the robot.
+        robot.startAuto()
+
+        // Grab the glyph.
+        val glyphGrabbingThread = thread(true) {
+            robot.closeGlyphGrabbers(1100)
+            thread(true) {
+                robot.lift.setPosition(RelicRecoveryRobotLift.LiftPosition.SECOND_LEVEL)
+            }
+        }
+
         // Start the timer.
         val elapsedTime = ElapsedTime(ElapsedTime.Resolution.MILLISECONDS)
-
-        // Start the robot.
-        robot.start()
-
-        robot.closeGlyphGrabbers(1200)
-        robot.lift.setPosition(RelicRecoveryRobotLift.LiftPosition.SECOND_LEVEL)
 
         // Find the position of the jewels.
         val jewelPosition = robot.jewelConfigurationDetector.waitForJewelIdentification(elapsedTime, this)
         robot.jewelConfigurationDetector.disable()
 
+        // Lower the jewel stick as soon as we determine the jewel configuration.
         if (jewelPosition != JewelConfigurationDetector.JewelConfiguration.UNKNOWN) {
             robot.lowerJewelStick(0)
         }
@@ -50,27 +50,38 @@ class RedFrontAutonomous : LinearOpMode() {
         val pictograph = pictographIdentifier.waitForPictographIdentification(elapsedTime, this)
         pictographIdentifier.deactivate()
 
+        // Wait for the glyph to be grabbed.
+        // (I find it funny that the robot identifies the jewels *and* pictograph
+        // before it even has time to grab the glyph that is right in front of it)
+        while (glyphGrabbingThread.isAlive) {
+            sleep(10)
+        }
+
         // Knock off the correct jewel.
         when (jewelPosition) {
             JewelConfigurationDetector.JewelConfiguration.RED_BLUE -> {
                 robot.timeDrive(1000, 0.175)
                 robot.raiseJewelStick()
-                robot.driveOnBalancingStone(-0.45)
+                robot.driveOnBalancingStone(-0.40)
                 robot.driveOffBalancingStone(-0.175)
+                robot.powerBreak(false)
             }
 
             JewelConfigurationDetector.JewelConfiguration.BLUE_RED -> {
                 robot.driveOffBalancingStone(-0.175)
+                robot.powerBreak(false)
                 robot.raiseJewelStick()
             }
 
             JewelConfigurationDetector.JewelConfiguration.UNKNOWN -> {
                 robot.driveOffBalancingStone(-0.175)
+                robot.powerBreak(false)
             }
         }
 
-        // Approach the crypto box.
-        robot.turn(0.45, -180.0)
+        thread(true) {
+            robot.lift.setPosition(RelicRecoveryRobotLift.LiftPosition.FIRST_LEVEL, 0.3)
+        }
 
         // Determine the distance from the wall to the correct crypto box.
         val rightWallDistance = when (pictograph) {
@@ -81,6 +92,9 @@ class RedFrontAutonomous : LinearOpMode() {
         }
 
         // Drive to the correct crypto box.
+        robot.timeDrive(750, 0.225)
+        robot.turn(0.3, 180.0)
+        //robot.timeDrive(750, -0.225)
         robot.driveToDistanceFromRightObject(rightWallDistance)
 
         // Place the glyph in the correct crypto box.
@@ -90,33 +104,42 @@ class RedFrontAutonomous : LinearOpMode() {
         robot.openGlyphGrabbers(250)
 
         // Back away from the crypto box.
-        robot.timeDrive(850, -0.175)
+        robot.timeDrive(1250, -0.225)
         robot.liftGlyphDeployer(500)
-
-        // Turn towards the center glyphs.
-        robot.turn(0.50, -45.0)
 
         // Open the glyph grabbers.
         robot.releaseGlyphGrabbers()
         robot.retractGlyphDeployer()
 
+        // Turn towards the center glyphs.
+        if (pictograph != RelicRecoveryVuMark.RIGHT) {
+            robot.driveToDistanceFromRightObject(RelicRecoveryRobot.TRAILING_FRONT_CRYPTO_BOX_DISTANCE, 1.00, false)
+        }
+
+        robot.turn(0.20, -45.0)
+
         // Drive to the glyphs in the center and grab one.
-        robot.timeDrive(2500)
+        sleep(1000)
+        robot.timeDrive(700, 1.0)
         robot.closeGlyphGrabbers(1250)
 
         // Drive back the crypto boxes.
-        robot.timeDrive(750, -0.25)
-        thread(true) { robot.lift.setPosition(RelicRecoveryRobotLift.LiftPosition.SECOND_LEVEL) }
-        robot.turn(0.75, 135.0)
+        robot.timeDrive(650, -0.25)
+        thread(true) { robot.lift.setPosition(RelicRecoveryRobotLift.LiftPosition.FIRST_LEVEL) }
+        robot.turn(0.35, 180.0)
         robot.timeDrive(500)
-        robot.turn(0.50, -180.0)
         robot.driveToDistanceFromForwardObject(RelicRecoveryRobot.CRYPTO_BOX_SPACING)
 
         // Line up with the center crypto box.
-        robot.driveToDistanceFromRightObject(RelicRecoveryRobot.CENTER_FRONT_CRYPTO_BOX_DISTANCE)
+        robot.driveToDistanceFromRightObject(RelicRecoveryRobot.TRAILING_FRONT_CRYPTO_BOX_DISTANCE)
 
         // Place the glyph in the correct crypto box.
-        robot.lift.drop()
+        if (pictograph == RelicRecoveryVuMark.LEFT) {
+            robot.lift.setPosition(RelicRecoveryRobotLift.LiftPosition.SECOND_LEVEL)
+        } else {
+            robot.lift.drop()
+        }
+
         robot.timeDrive(1000)
         robot.extendGlyphDeployer()
         robot.openGlyphGrabbers(250)
@@ -126,7 +149,9 @@ class RedFrontAutonomous : LinearOpMode() {
         robot.liftGlyphDeployer(500)
 
         // Turn towards the center glyphs.
-        robot.turn(0.25, -45.0)
+        robot.turn(0.35, -45.0)
+
+        */
     }
 
 }
